requirements.txt
streamlit
torch
pypdf
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub